requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
aiosignal==1.3.1
albumentations==1.1.0
attrs==23.1.0
cachetools==5.3.2
captum==0.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cycler==0.11.0
filelock==3.12.2
fonttools==4.43.0
frozenlist==1.4.0
fsspec==2023.6.0
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
huggingface-hub==0.20.1
idna==3.4
imageio==2.31.1
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
lazy_loader==0.3
lightning-utilities==0.9.0
matplotlib==3.5.2
minio==7.1.12
msgpack==1.0.5
networkx==3.1
numpy==1.22.4
onnx==1.15.0
onnxruntime-gpu==1.15.1
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.1
pandas==1.1.5
Pillow==10.0.1
pip==23.3.2
protobuf==4.24.0
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycryptodome==3.17
pyDeprecate==0.3.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
ray==2.6.2
referencing==0.30.2
regex==2023.8.8
reportlab==3.6.13
requests==2.31.0
rpds-py==0.9.2
rsa==4.9
safetensors==0.3.2
scikit-image==0.21.0
scikit-learn==1.1.1
scipy==1.11.1
seaborn==0.12.2
setuptools==68.0.0
simplejson==3.13.2
six==1.16.0
SimpleITK==2.3.1
tensorboardX==2.6.2
threadpoolctl==3.2.0
tifffile==2023.7.18
tokenizers==0.15.0
torch==2.0.1
torchmetrics==1.2.0
torchvision==0.15.2
tqdm==4.66.0
transformers==4.36.0
typing_extensions==4.7.1
urllib3==2.0.7
wheel==0.38.4
